frontend: refactor to make some controls available outside tables.
