+++ packages/frontend-web/src/app/scenes/Comments/components/ArticlePreview/ArticlePreview.tsx
@@ -35,13 +35,13 @@ import { ARTICLE_HEADER,
   VISUALLY_HIDDEN,
   WHITE_COLOR,
 } from '../../../../styles';
+import { abbreviateModerators } from '../../../../utilx';
 import { css, stylesheet } from '../../../../utilx';
 
 import {
   CloseIcon,
   OpenIcon,
 } from '../../../../components';
-import { abbreviateModerators } from '../../../../utilx';
 
 const STYLES = stylesheet({
   base: {
