+++ packages/frontend-web/src/app/scenes/Tables/styles.ts
@@ -23,29 +23,7 @@ import {
 } from '../../styles';
 import { stylesheet } from '../../utilx';
 
-export const IMAGE_BASE = 40;
 export const CELL_HEIGHT = 96;
-
-export const big = {
-  width: `${IMAGE_BASE}px`,
-  height: `${IMAGE_BASE}px`,
-};
-
-export const medium = {
-  width: `${IMAGE_BASE * 3 / 4}px`,
-  height: `${IMAGE_BASE * 3 / 4}px`,
-};
-
-export const small = {
-  width: `${IMAGE_BASE / 2}px`,
-  height: `${IMAGE_BASE / 2}px`,
-};
-
-export const flexCenter = {
-  display: 'flex',
-  alignItems: 'center',
-  justifyContent: 'center',
-};
 
 export const ARTICLE_TABLE_STYLES = stylesheet({
   dataTable: {
@@ -116,53 +94,5 @@ export const COMMON_STYLES = stylesheet({
     ':hover': {
       textDecoration: 'underline',
     },
-  },
-  smallIcon: {
-    width: `${IMAGE_BASE + 6}px`,
-    height: `${IMAGE_BASE + 6}px`,
-  },
-  xsmallIcon: {
-    width: `${IMAGE_BASE / 2 + 3}px`,
-    height: `${IMAGE_BASE / 2 + 3}px`,
-  },
-  smallImage: {
-    width: `${IMAGE_BASE}px`,
-    height: `${IMAGE_BASE}px`,
-    borderRadius: `${(IMAGE_BASE / 2)}px`,
-  },
-  xsmallImage: {
-    width: `${IMAGE_BASE / 2}px`,
-    height: `${IMAGE_BASE / 2}px`,
-    borderRadius: `${IMAGE_BASE / 4}px`,
-  },
-  textCenterSmall: {
-    ...small,
-    fontSize: '12px',
-    ...flexCenter,
-  },
-});
-
-export const ICON_STYLES = stylesheet({
-  big: big,
-  small: small,
-
-  iconCenter: {
-    width: `100%`,
-    height: `100%`,
-    ...flexCenter,
-  },
-
-  iconBackgroundCircle: {
-    ...big,
-    borderRadius: `${IMAGE_BASE}px`,
-    backgroundColor: '#eee',
-    display: 'inline-block',
-  },
-
-  iconBackgroundCircleSmall: {
-    ...small,
-    borderRadius: `${IMAGE_BASE / 2}px`,
-    backgroundColor: '#eee',
-    display: 'inline-block',
   },
 });
